+++ core/caiyun/options.ts
@@ -1,19 +1,19 @@
 export const defuConfig = {
-  shake: {
-    enable: true,
-    num: 15,
-    delay: 2,
+  "shake": {
+    "enable": true,
+    "num": 15,
+    "delay": 2
   },
-  garden: {
-    enable: true,
-    inviteCodes: [],
+  "garden": {
+    "enable": true,
+    "inviteCodes": []
   },
-  aiRedPack: {
-    enable: true,
+  "aiRedPack": {
+    "enable": true
   },
-  backupWaitTime: 20,
-  tasks: {
-    skipTasks: [],
+  "backupWaitTime": 20,
+  "tasks": {
+    "skipTasks": []
   },
-  catalog: '00019700101000000001',
+  "catalog": "00019700101000000001"
 }