+++ apps/caiyun/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @asunajs/caiyun
 
+## 0.3.5
+
+### Patch Changes
+
+- [`db9fd67`](https://github.com/yuiasn/asuna/commit/db9fd67f42a8ddf67417a27a79e67cadbf21753c) Thanks [@catlair](https://github.com/catlair)! - feat(caiyun): 输出版本信息
+
+- [`1d9b860`](https://github.com/yuiasn/asuna/commit/1d9b860fcade0a50a15383bbce12488edc980189) Thanks [@catlair](https://github.com/catlair)! - fix(caiyun): 打印任务完成情况
+
+## 0.3.4
+
+### Patch Changes
+
+- [`66bb167`](https://github.com/yuiasn/asuna/commit/66bb167d3a3e3212a4b63c206fa0dae0fd2ed613) Thanks [@catlair](https://github.com/catlair)! - feat(caiyun): 支持跳过任务
+
+## 0.3.3
+
+### Patch Changes
+
+- [`eda40a8`](https://github.com/yuiasn/asuna/commit/eda40a877e7be60f91f4c0c4651db29b99a4eda6) Thanks [@catlair](https://github.com/catlair)! - fix(caiyun): upload 1G task
+
 ## 0.3.2
 
 ### Patch Changes